red short cycling ! yay ! fix it.
