+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/State/Aura_AutoBlue_Short_State.java
@@ -33,7 +33,6 @@ import static com.qualcomm.robotcore.util.ElapsedTime.Resolution.MILLISECONDS;
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_1_RFI;
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_3_PS;
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_5_RFO_LOW;
-import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_5_RFO_MID;
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_6_PR_BOTH;
 import static org.firstinspires.ftc.teamcode.AuraRobot.APRILTAG_TIMEOUT;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_ACCEL_CONSTRAINT_1;
@@ -66,7 +65,6 @@ import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -87,7 +85,6 @@ import org.openftc.apriltag.AprilTagDetection;
 import org.openftc.easyopencv.OpenCvWebcam;
 
 import java.util.List;
-import java.util.Vector;
 import java.util.concurrent.TimeUnit;
 
 /*
@@ -135,7 +132,7 @@ public class Aura_AutoBlue_Short_State extends LinearOpMode {
 
     Vector2d blueReturnPos = new Vector2d(-40,57);
     Vector2d blueWingPos = new Vector2d(-60.75,57);
-    Vector2d blueStackPos = new Vector2d(-60.75,31.5);
+    Vector2d blueShortStackPos = new Vector2d(-60.75,31.5);
 
     Vector2d blueParkPos = new Vector2d(45, 54.5);  //7, 37
     boolean bProceedToYellow = false;
@@ -575,7 +572,7 @@ public class Aura_AutoBlue_Short_State extends LinearOpMode {
                 .strafeTo(blueEntryPos)
                 .stopAndAdd(rectifyHeadingError)
                 .strafeTo(blueWingPos, new TranslationalVelConstraint(65))
-                .strafeTo(blueStackPos)
+                .strafeTo(blueShortStackPos)
                 .afterDisp(0,deployStackIntake) // Make sure to turn on bottom roller
                 .stopAndAdd(intakeFromStack) // Make sure to flip box and lock fingers
                 .waitSeconds(AUTO_WAIT_FOR_STACK_INTAKE)
@@ -612,7 +609,7 @@ public class Aura_AutoBlue_Short_State extends LinearOpMode {
                 .strafeTo(blueEntryPos)
                 .stopAndAdd(rectifyHeadingError)
                 .strafeTo(blueWingPos, new TranslationalVelConstraint(65))
-                .strafeTo(blueStackPos)
+                .strafeTo(blueShortStackPos)
                 .afterDisp(0,deployStackIntake) // Make sure to turn on bottom roller
                 .stopAndAdd(intakeFromStack) // Make sure to flip box and lock fingers
                 .waitSeconds(AUTO_WAIT_FOR_STACK_INTAKE)
@@ -649,7 +646,7 @@ public class Aura_AutoBlue_Short_State extends LinearOpMode {
                 .strafeTo(blueEntryPos)
                 .stopAndAdd(rectifyHeadingError)
                 .strafeTo(blueWingPos, new TranslationalVelConstraint(65))
-                .strafeTo(blueStackPos)
+                .strafeTo(blueShortStackPos)
                 .afterDisp(0,deployStackIntake) // Make sure to turn on bottom roller
                 .stopAndAdd(intakeFromStack) // Make sure to flip box and lock fingers
                 .waitSeconds(AUTO_WAIT_FOR_STACK_INTAKE)
