+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/State/Aura_AutoRed_Short_State.java
@@ -36,10 +36,16 @@ import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState
 import static org.firstinspires.ftc.teamcode.AuraIntakeOuttakeController.ioState.STATE_6_PR_BOTH;
 import static org.firstinspires.ftc.teamcode.AuraRobot.APRILTAG_TIMEOUT;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_OUTTAKE;
+import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_PURPLE_DROP;
+import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_STACK_INTAKE;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_FOR_YELLOW_DROP;
 import static org.firstinspires.ftc.teamcode.AuraRobot.AUTO_WAIT_RETURN_TO_INTAKE;
 import static org.firstinspires.ftc.teamcode.AuraRobot.PURPLE_LOCK;
 import static org.firstinspires.ftc.teamcode.AuraRobot.PURPLE_UNLOCK;
+import static org.firstinspires.ftc.teamcode.AuraRobot.leftLinkageClose;
+import static org.firstinspires.ftc.teamcode.AuraRobot.leftLinkageOpen;
+import static org.firstinspires.ftc.teamcode.AuraRobot.rightLinkageClose;
+import static org.firstinspires.ftc.teamcode.AuraRobot.rightLinkageOpen;
 
 import android.util.Size;
 
@@ -51,6 +57,7 @@ import com.acmerobotics.roadrunner.Action;
 import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.SequentialAction;
+import com.acmerobotics.roadrunner.TranslationalVelConstraint;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
@@ -95,31 +102,37 @@ import java.util.concurrent.TimeUnit;
 
 public class Aura_AutoRed_Short_State extends LinearOpMode {
 
-    //Todo:switch to field coordinates, x and heading inverse of Red Short
-
-
     //**** Roadrunner Pose2ds ****
 
+    //Todo: switch to field coordinates:
+    //to find x: add robotcentric Y pos to field centric start pos X
+    //to find y: subtract robotcentric X pos to field centric start pos y
+    //to find heading: add -90 degrees to field centric start pos heading
+
+
+    // RObot Width = 15; Length = 15.5
     Pose2d redStartPos = new Pose2d(15,-61.5,Math.toRadians(90));//0,0,0
 
-    Pose2d redPurple3Pos = new Pose2d(14, -33, Math.toRadians(0)); //27,19,-90
+    Pose2d redPurple3Pos = new Pose2d(14, -33, Math.toRadians(0));  //27,0,-90
     Pose2d redPurple2Pos = new Pose2d(16, -32, Math.toRadians(90));  //37,12,-90
-    Pose2d redPurple1Pos = new Pose2d(8, -33, Math.toRadians(180));  //27,0,-90
+    Pose2d redPurple1Pos = new Pose2d(8, -33 , Math.toRadians(180)); //27,19,-90
 
     Pose2d redTagPos = new Pose2d(36,-45, Math.toRadians(0));
 
     Vector2d redYellow3Pos = new Vector2d(50.5, -42);  //27,37,-90
-    Vector2d redYellow2Pos = new Vector2d(50.5, -32);   //26,37,-90
-    Vector2d redYellow1Pos = new Vector2d(50.5,-27.5);    //33,37,-90
+    Vector2d redYellow2Pos = new Vector2d(50.5, -34);   //26,37,-90
+    Vector2d redYellow1Pos = new Vector2d(51.5,-28);    //33,37,-90
 
-    Vector2d redParkPos = new Vector2d(45, -54.5);  //7, 37
+    Pose2d redBeforeGateCyclePos = new Pose2d(12,-57, Math.toRadians(0));
 
     Vector2d redEntryPos = new Vector2d(12,-57);
 
-    Vector2d blueReturnPos = new Vector2d(-40,-57);
-    Vector2d blueWingPos = new Vector2d(-60.75,-57);
-    Vector2d blueStackPos = new Vector2d(-60.75,-31.5);
+    Vector2d redReturnPos = new Vector2d(-40,-57);
+    Vector2d redWingPos = new Vector2d(-60.75,-57);
+    Vector2d redStackPos = new Vector2d(-60.75,-32.5);
 
+    Vector2d redParkPos = new Vector2d(45, -54.5);  //7, 37
+    boolean bProceedToYellow = false;
 
     //Roadrunner field-centric coordinates quick guide brought to you by Lavanya
 
@@ -169,6 +182,7 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
     }
     public Action getReadyForOutTake = new GotoOuttakeAction();
 
+
     public class GotoIntakeAction implements Action {
         @Override
         public boolean run(TelemetryPacket tPkt) {
@@ -202,24 +216,80 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
     }
     public Action endTrajectoryMarker = new endTrajectory();
 
-    public class backwallAprilTagController implements Action {
+    public class backwallAprilTagControllerEyeball implements Action {
         @Override
         public boolean run(TelemetryPacket tPkt) {
-            updatePosfromBackwallAprilTag();
+            updateposFromAprilTagEyeball();
             return false;
         }
     }
 
-    public Action updateAfterGatePos = new backwallAprilTagController();
+    public class backwallAprilTagControllerKemera implements Action {
+        @Override
+        public boolean run(TelemetryPacket tPkt) {
+            updateposFromAprilTagKemera();
+            return false;
+        }
+    }
+
+
+    public Action updatePosFromAprilTagEyeball = new backwallAprilTagControllerEyeball();
+    public Action updatePosFromAprilTagKemera = new backwallAprilTagControllerKemera();
+
 
     public class initAprilTag implements Action {
         public boolean run(TelemetryPacket tPkt){
-            initAprilTag();
+//            initAprilTagEyeball();
+            initAprilTagKemera();
             return false;
         }
     }
 
     public Action initApril = new initAprilTag();
+
+
+    public class lowerStackLinkage implements Action {
+        public boolean run(TelemetryPacket tPkt){
+            lowerStackIntake();
+            return false;
+        }
+    }
+    public Action deployStackIntake = new lowerStackLinkage();
+
+
+    public class raiseStackLinkage implements Action {
+        public boolean run(TelemetryPacket tPkt){
+            raiseStackIntake();
+            return false;
+        }
+    }
+    public Action retractStackIntake = new raiseStackLinkage();
+
+    public class stackIntake implements Action {
+        public boolean run(TelemetryPacket tPkt){
+            stackIntakePixels();
+            return false;
+        }
+    }
+    public Action intakeFromStack = new stackIntake();
+
+    public class ReverseIntake implements Action {
+        public boolean run(TelemetryPacket tPkt){
+            reverseIntake();
+            return false;
+        }
+    }
+    public Action reverseInt = new ReverseIntake();
+
+
+    public class raiseBox implements Action {
+        public boolean run(TelemetryPacket tPkt){
+            secureStackIntakePixels();
+            return false;
+        }
+    }
+    public Action securePixels = new raiseBox();
+
 
     private static final double LEFT_SPIKEMARK_BOUNDARY_X = 300;
     private static final double RIGHT_SPIKEMARK_BOUNDARY_X = 130;
@@ -249,7 +319,7 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
             telemetry.addData("Corrected heading:", Math.toDegrees(yaw));
             telemetry.update();
 
-           RedShort.pose = new Pose2d(RedShort.pose.position.x,RedShort.pose.position.y, yaw);
+            RedShort.pose = new Pose2d(RedShort.pose.position.x, RedShort.pose.position.y, yaw);
 
             return false;
         }
@@ -296,7 +366,7 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
 
     // TFOD_MODEL_FILE points to a model file stored onboard the Robot Controller's storage,
     // this is used when uploading models directly to the RC using the model upload interface.
-//    private static final String TFOD_MODEL_FILE = "C:\\Sashank\\FTC CenterStage\\Aurelius\\Aurelius\\TeamCode\\src\\main\\java\\org\\firstinspires\\ftc\\teamcode\\myBloopy.tflite";
+//    private static final String TFOD_MODEL_FILE = "C:\\Sashank\\FTC CenterStage\\Aurelius\\Aurelius\\TeamCode\\src\\main\\java\\org\\firstinspires\\ftc\\teamcode\\myRedpy.tflite";
 
     // Define the labels recognized in the model for TFOD (must be in training order!)
     private static final String[] LABELS = {
@@ -310,11 +380,14 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
     /**
      * The variable to store our instance of the vision portal.
      */
-    private VisionPortal visionPortal;
+    private VisionPortal visionPortalEyeball;
+    private VisionPortal visionPortalKemera;
+    private VisionPortal visionPortalTfod;
 
     private static final boolean USE_WEBCAM = true;
     public static final int DESIRED_TAG_ID = -1;     // Choose the tag you want to approach or set to -1 for ANY tag.
-    private AprilTagProcessor aprilTag;              // Used for managing the AprilTag detection process.
+    private AprilTagProcessor aprilTagEyeball;              // Used for managing the AprilTag detection process.
+    private AprilTagProcessor aprilTagKemera;
     private org.firstinspires.ftc.vision.apriltag.AprilTagDetection desiredTag = null;     // Used to hold the data for a detected AprilTag
     boolean targetFound     = false;    // Set to true when an AprilTag target is detected
 
@@ -347,6 +420,8 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
         Aurelius.PurpleDumper.setPosition(PURPLE_LOCK);
         Aurelius.boeing747.init();
         Aurelius.hanger.init();
+        Aurelius.LeftLink.setPosition(leftLinkageClose);
+        Aurelius.RightLink.setPosition(rightLinkageClose);
         Aurelius.hanger.update();
         telemetry.addLine(String.format("%d. Aura Initialized!", iTeleCt++));
         telemetry.update();
@@ -377,6 +452,9 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
         telemetry.addData("Status: ", "Initializing Tensor Flow ......");
         telemetry.update();
         initTfod();
+        //initAprilTagEyeball();
+        //initAprilTagKemera();
+
         runtime.reset();
         while(runtime.seconds() < 3)
             idle();
@@ -400,7 +478,7 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
         runtime.reset();
         if (opModeIsActive()) {
             DetectPurpleDropoffPos();
-            visionPortal.close();
+            visionPortalKemera.close();
             telemetry.addData("Going to position:", "PurpleDropOffPos");
             telemetry.update();
 
@@ -408,43 +486,43 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
             switch (PurpleDropOffPos) {
                 case 1:
                     Actions.runBlocking(
-                        new ParallelAction(
-                            new SequentialAction(
-                                beginTrajectoryMarker,
-                                dropOffPurpleAtPos1,
-                                dropOffYellowAtPos1,
-                                endTrajectoryMarker),
-                                new ParallelAction(
-                                        updateIOController,
-                                        initApril)
-                        ));
+                            new ParallelAction(
+                                    new SequentialAction(
+                                            beginTrajectoryMarker,
+                                            dropOffPurpleAtPos1,
+                                            dropOffYellowAtPos1,
+                                            endTrajectoryMarker),
+                                    new ParallelAction(
+                                            updateIOController,
+                                            initApril)
+                            ));
                     break;
                 case 2:
                     Actions.runBlocking(
-                    new ParallelAction(
-                        new SequentialAction(
-                            beginTrajectoryMarker,
-                            dropOffPurpleAtPos2,
-                            dropOffYellowAtPos2,
-                            endTrajectoryMarker),
                             new ParallelAction(
-                                    updateIOController,
-                                    initApril)
-                    ));
+                                    new SequentialAction(
+                                            beginTrajectoryMarker,
+                                            dropOffPurpleAtPos2,
+                                            dropOffYellowAtPos2,
+                                            endTrajectoryMarker),
+                                    new ParallelAction(
+                                            updateIOController,
+                                            initApril)
+                            ));
                     break;
                 case 3:
                 default:
                     Actions.runBlocking(
-                    new ParallelAction(
-                        new SequentialAction(
-                            beginTrajectoryMarker,
-                            dropOffPurpleAtPos3,
-                            dropOffYellowAtPos3,
-                            endTrajectoryMarker),
                             new ParallelAction(
-                                    updateIOController,
-                                    initApril)
-                    ));
+                                    new SequentialAction(
+                                            beginTrajectoryMarker,
+                                            dropOffPurpleAtPos3,
+                                            dropOffYellowAtPos3,
+                                            endTrajectoryMarker),
+                                    new ParallelAction(
+                                            updateIOController,
+                                            initApril)
+                            ));
                     break;
             }
         }
@@ -456,21 +534,21 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
                 .setTangent(Math.toRadians(0))
                 .splineToLinearHeading(redPurple3Pos, Math.toRadians(90))
                 .stopAndAdd(ejectPurple)
-                .waitSeconds(1)
+                .waitSeconds(AUTO_WAIT_FOR_PURPLE_DROP)
                 .build();
 
         dropOffPurpleAtPos2 = RedShort.actionBuilder(redStartPos)
                 .setTangent(Math.toRadians(0))
                 .splineToLinearHeading(redPurple2Pos, Math.toRadians(90))
                 .stopAndAdd(ejectPurple)
-                .waitSeconds(1)
+                .waitSeconds(AUTO_WAIT_FOR_PURPLE_DROP)
                 .build();
 
         dropOffPurpleAtPos1 = RedShort.actionBuilder(redStartPos)
                 .setTangent(Math.toRadians(80))
                 .splineToLinearHeading(redPurple1Pos, Math.toRadians(130))
                 .stopAndAdd(ejectPurple)
-                .waitSeconds(1)
+                .waitSeconds(AUTO_WAIT_FOR_PURPLE_DROP)
                 .build();
     }
 
@@ -482,8 +560,29 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
                 .setTangent(Math.toRadians(-90))
                 .splineToLinearHeading(redTagPos, Math.toRadians(90))
                 .afterDisp(0, getReadyForOutTake)
-                .stopAndAdd(updateAfterGatePos)
+                .stopAndAdd(updatePosFromAprilTagKemera)
                 .strafeTo(redYellow3Pos)
+                .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
+                .stopAndAdd(depositYellow)
+                .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
+                .afterDisp(0, getReadyForIntake)
+                .strafeTo(redEntryPos)
+                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redWingPos, new TranslationalVelConstraint(65))
+                .strafeTo(redStackPos)
+                .afterDisp(0,deployStackIntake) // Make sure to turn on bottom roller
+                .stopAndAdd(intakeFromStack) // Make sure to flip box and lock fingers
+                .waitSeconds(AUTO_WAIT_FOR_STACK_INTAKE)
+                .stopAndAdd(securePixels)
+                .strafeTo(redReturnPos)
+                .afterDisp(0, retractStackIntake)
+//                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redEntryPos)
+                .afterDisp(50, getReadyForOutTake)
+                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redTagPos.position)
+                .stopAndAdd(updatePosFromAprilTagKemera)
+                .strafeTo(redYellow2Pos)
                 .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
                 .stopAndAdd(depositYellow)
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
@@ -494,13 +593,32 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
 
         dropOffYellowAtPos2 = RedShort.actionBuilder(redPurple2Pos)
                 .stopAndAdd(rectifyHeadingError)
-                .lineToY(-44.5)
+                .strafeTo(new Vector2d(10,-38.5))
                 .setTangent(Math.toRadians(-90))
                 .splineToLinearHeading(redTagPos, Math.toRadians(90))
                 .afterDisp(0, getReadyForOutTake)
-                .stopAndAdd(updateAfterGatePos)
-                .strafeTo(redYellow3Pos)
+                .stopAndAdd(updatePosFromAprilTagKemera)
                 .strafeTo(redYellow2Pos)
+                .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
+                .stopAndAdd(depositYellow)
+                .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
+                .afterDisp(0, getReadyForIntake)
+                .strafeTo(redEntryPos)
+                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redWingPos, new TranslationalVelConstraint(65))
+                .strafeTo(redStackPos)
+                .afterDisp(0,deployStackIntake) // Make sure to turn on bottom roller
+                .stopAndAdd(intakeFromStack) // Make sure to flip box and lock fingers
+                .waitSeconds(AUTO_WAIT_FOR_STACK_INTAKE)
+                .stopAndAdd(securePixels)
+                .strafeTo(redReturnPos)
+                .afterDisp(0, retractStackIntake)
+                .strafeTo(redEntryPos)
+                .afterDisp(50, getReadyForOutTake)
+                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redTagPos.position)
+                .stopAndAdd(updatePosFromAprilTagKemera)
+                .strafeTo(redYellow1Pos)
                 .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
                 .stopAndAdd(depositYellow)
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
@@ -509,15 +627,35 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
                 .waitSeconds(AUTO_WAIT_RETURN_TO_INTAKE)
                 .build();
 
+
         dropOffYellowAtPos1 = RedShort.actionBuilder(redPurple1Pos)
                 .stopAndAdd(rectifyHeadingError)
-                .lineToX(14)
+                .strafeTo(new Vector2d(10,-38.5))
                 .setTangent(Math.toRadians(-90))
                 .splineToLinearHeading(redTagPos, Math.toRadians(90))
                 .afterDisp(0, getReadyForOutTake)
-                .stopAndAdd(updateAfterGatePos)
-                .strafeTo(redYellow3Pos)
+                .stopAndAdd(updatePosFromAprilTagKemera)
                 .strafeTo(redYellow1Pos)
+                .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
+                .stopAndAdd(depositYellow)
+                .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
+                .afterDisp(0, getReadyForIntake)
+                .strafeTo(redEntryPos)
+                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redWingPos, new TranslationalVelConstraint(65))
+                .strafeTo(redStackPos)
+                .afterDisp(0,deployStackIntake) // Make sure to turn on bottom roller
+                .stopAndAdd(intakeFromStack) // Make sure to flip box and lock fingers
+                .waitSeconds(AUTO_WAIT_FOR_STACK_INTAKE)
+                .stopAndAdd(securePixels)
+                .strafeTo(redReturnPos)
+                .afterDisp(0, retractStackIntake)
+                .strafeTo(redEntryPos)
+                .afterDisp(50, getReadyForOutTake)
+                .stopAndAdd(rectifyHeadingError)
+                .strafeTo(redTagPos.position)
+                .stopAndAdd(updatePosFromAprilTagKemera)
+                .strafeTo(redYellow2Pos)
                 .waitSeconds(AUTO_WAIT_FOR_OUTTAKE)
                 .stopAndAdd(depositYellow)
                 .waitSeconds(AUTO_WAIT_FOR_YELLOW_DROP)
@@ -561,7 +699,7 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
         builder.addProcessor(tfod);
 
         // Build the Vision Portal, using the above settings.
-        visionPortal = builder.build();
+        visionPortalKemera = builder.build();
 
     } // end method initTfod()
 
@@ -612,17 +750,156 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
 
     }
 
-    boolean updatePosfromBackwallAprilTag()
+    boolean updateposFromAprilTagEyeball()
     {
 //        initAprilTag(); // initializing the april tag processor
-        setManualExposure(6, 250); // accounting for motion blur
+        setManualExposureEyeball(6, 250); // accounting for motion blur
         targetFound = false;
         desiredTag  = null;
 
         ElapsedTime AprilTagTimer = new ElapsedTime();
         AprilTagTimer.reset();
         while(!targetFound && AprilTagTimer.seconds() < APRILTAG_TIMEOUT) {
-            List<org.firstinspires.ftc.vision.apriltag.AprilTagDetection> currentDetections = aprilTag.getDetections();
+            List<org.firstinspires.ftc.vision.apriltag.AprilTagDetection> currentDetections = aprilTagEyeball.getDetections();
+            for (org.firstinspires.ftc.vision.apriltag.AprilTagDetection detection : currentDetections) {
+                // Look to see if we have size info on this tag.
+                if (detection.metadata != null) {
+                    //  Check to see if we want to track towards this tag.
+                    if ((DESIRED_TAG_ID < 0) || (detection.id == DESIRED_TAG_ID)) {
+                        targetFound = true;
+                        desiredTag = detection;
+                        break;
+                    } else {
+                        telemetry.addData("Skipping", "Tag ID %d is not desired", detection.id);
+                        telemetry.update();
+                    }
+                } else {
+                    telemetry.addData("Unknown", "Tag ID %d is not in TagLibrary", detection.id);
+                    telemetry.update();
+                }
+            }
+        }
+
+        if(targetFound) {
+            telemetry.addData("Found", "ID %d (%s)", desiredTag.id, desiredTag.metadata.fieldPosition);
+            telemetry.addData("Range",  "%5.1f inches", desiredTag.ftcPose.range);
+            telemetry.addData("Bearing","%3.0f degrees", desiredTag.ftcPose.bearing);
+            telemetry.addData("Yaw","%3.0f degrees", desiredTag.ftcPose.yaw);
+            telemetry.update();
+
+            // TODO: 1. Calibrate camera
+            //       2. Offset currX and currY from camera to robot center
+
+            double range = desiredTag.ftcPose.range;
+            double bearing = desiredTag.ftcPose.bearing;
+            double yaw = desiredTag.ftcPose.yaw;
+
+            double robotOffsetX = -7;
+            double robotOffsetY = +5.5;
+
+            double offsetX = (range * Math.cos(Math.toRadians(bearing)));
+
+            double offsetY = (range * Math.sin(Math.toRadians(bearing)));
+
+            double currHeading = -Math.toRadians(yaw);
+
+            double rotateX = (robotOffsetX * Math.cos(currHeading)) + (robotOffsetY * -Math.sin(currHeading));
+            double rotateY = (robotOffsetX * Math.sin(currHeading)) + (robotOffsetY * Math.cos(currHeading));
+
+            double currX = rotateX + (desiredTag.metadata.fieldPosition.getData()[0] -
+                    offsetX);
+
+            double currY = rotateY + (desiredTag.metadata.fieldPosition.getData()[1] -
+                    offsetY);
+
+            telemetry.addData("Current pos:", "X: %5.1f Y: %5.1f Heading: %5.1f degrees", RedShort.pose.position.x, RedShort.pose.position.y, Math.toDegrees(RedShort.pose.heading.log()));
+            telemetry.update();
+
+            RedShort.pose = new Pose2d(currX, currY, currHeading);
+            telemetry.addData("Updated pos:", "X: %5.1f Y: %5.1f Heading %5.1f degrees", RedShort.pose.position.x, RedShort.pose.position.y, Math.toDegrees(RedShort.pose.heading.log()));
+            telemetry.update();
+            return true;
+        }
+        telemetry.addLine("Not Found: Desired Tag not found");
+        telemetry.update();
+        return false;
+    }
+
+    private void initAprilTagEyeball() {
+        // Create the AprilTag processor by using a builder.
+        aprilTagEyeball = new AprilTagProcessor.Builder()
+                .setLensIntrinsics(822.317f, 822.317f, 319.495f, 242.502f)
+                .build();
+
+        // Adjust Image Decimation to trade-off detection-range for detection-rate.
+        // eg: Some typical detection data using a Logitech C920 WebCam
+        // Decimation = 1 ..  Detect 2" Tag from 10 feet away at 10 Frames per second
+        // Decimation = 2 ..  Detect 2" Tag from 6  feet away at 22 Frames per second
+        // Decimation = 3 ..  Detect 2" Tag from 4  feet away at 30 Frames Per Second
+        // Decimation = 3 ..  Detect 5" Tag from 10 feet away at 30 Frames Per Second
+        // Note: Decimation can be changed on-the-fly to adapt during a match.
+        aprilTagEyeball.setDecimation(2);
+
+        // Create the vision portal by using a builder.
+        if (USE_WEBCAM) {
+            visionPortalEyeball = new VisionPortal.Builder()
+                    .setCamera(hardwareMap.get(WebcamName.class, "Eyeball"))
+                    .setCameraResolution(new Size(640, 480))
+                    .addProcessor(aprilTagEyeball)
+                    .build();
+        } else {
+            visionPortalEyeball = new VisionPortal.Builder()
+                    .setCamera(BuiltinCameraDirection.BACK)
+                    .addProcessor(aprilTagEyeball)
+                    .build();
+        }
+    }
+
+    private void setManualExposureEyeball(int exposureMS, int gain) {
+        // Wait for the camera to be open, then use the controls
+
+        if (visionPortalEyeball == null) {
+            return;
+        }
+
+        // Make sure camera is streaming before we try to set the exposure controls
+        if (visionPortalEyeball.getCameraState() != VisionPortal.CameraState.STREAMING) {
+            telemetry.addData("Camera", "Waiting");
+            telemetry.update();
+            while (!isStopRequested() && (visionPortalEyeball.getCameraState() != VisionPortal.CameraState.STREAMING)) {
+                sleep(20);
+            }
+            telemetry.addData("Camera", "Ready");
+            telemetry.update();
+        }
+
+        // Set camera controls unless we are stopping.
+        if (!isStopRequested())
+        {
+            ExposureControl exposureControl = visionPortalEyeball.getCameraControl(ExposureControl.class);
+            if (exposureControl.getMode() != ExposureControl.Mode.Manual) {
+                exposureControl.setMode(ExposureControl.Mode.Manual);
+                sleep(50);
+            }
+            exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
+            sleep(20);
+            GainControl gainControl = visionPortalEyeball.getCameraControl(GainControl.class);
+            gainControl.setGain(gain);
+            sleep(20);
+        }
+    }
+
+    boolean updateposFromAprilTagKemera()
+    {
+//        initAprilTag(); // initializing the april tag processor
+        setManualExposureKemera(6, 250); // accounting for motion blur
+        targetFound = false;
+        desiredTag  = null;
+
+        ElapsedTime AprilTagTimer = new ElapsedTime();
+        AprilTagTimer.reset();
+        while(!targetFound && AprilTagTimer.seconds() < APRILTAG_TIMEOUT) {
+            List<org.firstinspires.ftc.vision.apriltag.AprilTagDetection> currentDetections = aprilTagKemera.getDetections();
             for (org.firstinspires.ftc.vision.apriltag.AprilTagDetection detection : currentDetections) {
                 // Look to see if we have size info on this tag.
                 if (detection.metadata != null) {
@@ -687,11 +964,12 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
         return false;
     }
 
-    private void initAprilTag() {
+    private void initAprilTagKemera() {
         // Create the AprilTag processor by using a builder.
-        aprilTag = new AprilTagProcessor.Builder()
+        aprilTagKemera = new AprilTagProcessor.Builder()
                 .setLensIntrinsics(822.317f, 822.317f, 319.495f, 242.502f)
                 .build();
+
         // Adjust Image Decimation to trade-off detection-range for detection-rate.
         // eg: Some typical detection data using a Logitech C920 WebCam
         // Decimation = 1 ..  Detect 2" Tag from 10 feet away at 10 Frames per second
@@ -699,35 +977,35 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
         // Decimation = 3 ..  Detect 2" Tag from 4  feet away at 30 Frames Per Second
         // Decimation = 3 ..  Detect 5" Tag from 10 feet away at 30 Frames Per Second
         // Note: Decimation can be changed on-the-fly to adapt during a match.
-        aprilTag.setDecimation(2);
+        aprilTagKemera.setDecimation(2);
 
         // Create the vision portal by using a builder.
         if (USE_WEBCAM) {
-                visionPortal = new VisionPortal.Builder()
-                        .setCamera(hardwareMap.get(WebcamName.class, "Kemera"))
-                        .setCameraResolution(new Size(640, 480))
-                        .addProcessor(aprilTag)
-                        .build();
+            visionPortalKemera = new VisionPortal.Builder()
+                    .setCamera(hardwareMap.get(WebcamName.class, "Kemera"))
+                    .setCameraResolution(new Size(640, 480))
+                    .addProcessor(aprilTagKemera)
+                    .build();
         } else {
-            visionPortal = new VisionPortal.Builder()
+            visionPortalKemera = new VisionPortal.Builder()
                     .setCamera(BuiltinCameraDirection.BACK)
-                    .addProcessor(aprilTag)
+                    .addProcessor(aprilTagKemera)
                     .build();
         }
     }
 
-    private void setManualExposure(int exposureMS, int gain) {
+    private void setManualExposureKemera(int exposureMS, int gain) {
         // Wait for the camera to be open, then use the controls
 
-        if (visionPortal == null) {
+        if (visionPortalKemera == null) {
             return;
         }
 
         // Make sure camera is streaming before we try to set the exposure controls
-        if (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING) {
+        if (visionPortalKemera.getCameraState() != VisionPortal.CameraState.STREAMING) {
             telemetry.addData("Camera", "Waiting");
             telemetry.update();
-            while (!isStopRequested() && (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING)) {
+            while (!isStopRequested() && (visionPortalKemera.getCameraState() != VisionPortal.CameraState.STREAMING)) {
                 sleep(20);
             }
             telemetry.addData("Camera", "Ready");
@@ -737,17 +1015,47 @@ public class Aura_AutoRed_Short_State extends LinearOpMode {
         // Set camera controls unless we are stopping.
         if (!isStopRequested())
         {
-            ExposureControl exposureControl = visionPortal.getCameraControl(ExposureControl.class);
+            ExposureControl exposureControl = visionPortalKemera.getCameraControl(ExposureControl.class);
             if (exposureControl.getMode() != ExposureControl.Mode.Manual) {
                 exposureControl.setMode(ExposureControl.Mode.Manual);
                 sleep(50);
             }
             exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
             sleep(20);
-            GainControl gainControl = visionPortal.getCameraControl(GainControl.class);
+            GainControl gainControl = visionPortalKemera.getCameraControl(GainControl.class);
             gainControl.setGain(gain);
             sleep(20);
         }
+    }
+
+    public void reverseIntake()
+    {
+        Aurelius.setPower(AuraRobot.AuraMotors.INTAKE, -.7);
+    }
+
+    public void lowerStackIntake()
+    {
+        Aurelius.LeftLink.setPosition(leftLinkageOpen);
+        Aurelius.RightLink.setPosition(rightLinkageOpen);
+        Aurelius.Roller.setPower(-0.7);
+    }
+
+    public void raiseStackIntake()
+    {
+        Aurelius.LeftLink.setPosition(leftLinkageClose);
+        Aurelius.RightLink.setPosition(rightLinkageClose);
+        Aurelius.Roller.setPower(0);
+    }
+
+    public void stackIntakePixels()
+    {
+        Aurelius.setPower(AuraRobot.AuraMotors.INTAKE, 0.7);
+    }
+
+    public void secureStackIntakePixels()
+    {
+        Aurelius.setPower(AuraRobot.AuraMotors.INTAKE, 0);
+        MyIntakeOuttakeController.setTargetState(STATE_3_PS);
     }
